Update my_panel.py
This is the second batch of changes for this branch. Added a drilldown feature: when an assorted match is picked and it is part of the search term, that item will be searched and show in the quick_panel, whereas the original text in the input_panel will be maintained.