Update my_panel.py
Forgot to cater for "case_i" when i added the code for "visual effects" on March 20. Besides, i also forgot to reset the "orisel" after any text is picked and inserted into the buffer, which should have been part of the code added for, again, "visual effects" on March 20. so here you go, this "commit" have them put on. Let's refer this "commit" as  "23h23h1" (coz i will eventually refer it in another branch.)