Update my_panel.py
Some of the behaviors between "view.find_all" and "re.findall" are inconsistent (perhaps they adopted different regular expression standards) enough that i decided to avoid "re.findall" (even though it is rather handy) and use only "view.find_all" to achieve everything i needed. so, this "commit" is all about it, as well as a minor code adjustment (to move some code out of the loop, as they were not needed inside, but before it). that's all.