Update my_panel.py
Added a mechanism, which slimed the search result, to improve response time when too many hits have been returned from the search. The final outcome is not much changed while the response time is improved. A slight bump is nevertheless caused. Some of the existing codes have been tuned to work with it. Let's refer this "commit" as "23h24F1" (coz i will eventually refer it in another branch.)