Update my_panel.py
It used to be a two-level arrangement in terms of user interface. That was, it always show_input_panel (for input) first and therefore show_quick_panel (for search result, or history list if null string is entered). Now, to simplify (in operational point of view) it a bit, it goes straight to the history list if no selection is found for search. (So, this also implied if a selection is found for search, it will go straight to search and show you the search result.) You can select an item and edit upon, or type your search term directly there (since the history list will keep on shortening as you type, you can always pick one from the list). When you finish your typing and it's non-existent (i.e. nothing matched on the list), you have to type a double-semicolon instead of hitting the enter key though. Because the enter key is not functioning when nothing matched. In fact, please note, the enter key there is NOT for ending your typing and taking whatever you have typed for search. It's only for selecting the highlighted item. So, by all means, the safest bet there, is always use a double-semicolon to end it, unless you eventually go for picking an item from the list. By the way, the double-semicolon trick also works in active_view, it simply triggers the search (or a history list if the "##something" (in which, "something" means you have something after "##" and before where you typed a double-semicolon) is not found) there.