Update my_panel.py
Added a popup (which serves as an action button) in history list for item-delete, as a better user interface for the job than the existing one. Since the popup inevitably obstructed the first row of the list, i added a dummy item on top of the list to minimize the impact. However, as soon as you start typing, the list gets shortening that obstruction will still be encountered. To get around it, i suggest you type a few tab characters beforehand. Because, the popup will stick with the caret hence a few tabs will move it to the right for certain distance and make the text underneath visible again.