Update my_panel.py
Changed the method to capture a double-semicolon, so that, only those consecutive semicolons within two seconds will be recognized as a valid double-semicolon. This way can increase the flexibility in text typing when `;;` is deliberately needed, simply by waiting for a while after the first `;` and before the second `;`. Otherwise, you may have to make use of `; ;` `leftarrow key` `backspace` `rightarrow key` or `;` `leftarrow key` `;` `rightarrow key` those kinds of hastles to get it around.