Update my_panel.py
imesh (ime shorthand) search is a new feature which will take the content of the current buffer and the udmtable to generate an imesh searchable version of the content in memory hence user can search text by any improvised ime shorthands. since generating the imesh searchable content takes time, i figured to keep it in memory for subsequent searches until user issues a "renew" instruction to regenerate it. i included a sample text file (which has lyrics) to show the power of this new feature. just open it up in sublime text editor and search this `/[[sws]]/=np=` to get a sense of it. to renew the content in memory, just add `[r]` or `[rn]` or `[renew]` in your search term, for instance, `/[[yjy]]/=np=[r]`.