Update my_panel.py
added a feature for reverse order. it could be useful in dealing with a big log file that you very often interested in the matches close to the bottom than to the top. add this `=rv=` in your search term to get the search results listed in reverse order (by line).