Update my_panel.py
added a feature to load last result back on screen. it can save time when your file is big and/or your search term is sophisticated which takes noticeable time to complete, "load last result" is handy to show the last search result in memory without having to search again. enter this `[ll]` as your search term (or it exists anywhere in your search term) to get the result back.