Update my_panel.py
i found that the coverage of the erasure of the "skip" setting was incomplete. hence, adjusted it by relocating them from the bottom line of the concerned functions to the very beginning there.

i also added two more menu items, they are, `Show History List of Useful Search` and `Force Shorthand Search`.